Rename and move ReferenceDataRenderer to LabelRenderer
