+++ src/Pim/Bundle/ReferenceDataBundle/MongoDB/Normalizer/ReferenceDataNormalizer.php
@@ -2,27 +2,28 @@
 
 namespace Pim\Bundle\ReferenceDataBundle\MongoDB\Normalizer;
 
-use Pim\Bundle\ReferenceDataBundle\DataGrid\ReferenceDataRenderer;
+use Pim\Component\ReferenceData\LabelRenderer;
 use Pim\Component\ReferenceData\Model\ReferenceDataInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
  * Normalize a reference data very simply to store it as mongodb_json
+ *
  * @author    Julien Janvier <julien.janvier@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class ReferenceDataNormalizer implements NormalizerInterface
 {
-    /** @var ReferenceDataRenderer */
+    /** @var LabelRenderer */
     protected $renderer;
 
     /**
      * Constructor
      *
-    *@param ReferenceDataRenderer $renderer
+     * @param LabelRenderer $renderer
      */
-    public function __construct(ReferenceDataRenderer $renderer)
+    public function __construct(LabelRenderer $renderer)
     {
         $this->renderer = $renderer;
     }
