+++ src/Pim/Bundle/ReferenceDataBundle/Resources/config/services.yml
@@ -3,7 +3,7 @@ parameters:
     pim_reference_data.method_guesser.class: Pim\Component\ReferenceData\MethodNameGuesser
     pim_reference_data.repository_resolver.class: Pim\Bundle\ReferenceDataBundle\Doctrine\ReferenceDataRepositoryResolver
     pim_reference_data.id_resolver.class: Pim\Bundle\ReferenceDataBundle\Doctrine\ReferenceDataIdResolver
-    pim_reference_data.datagrid_renderer.class: Pim\Bundle\ReferenceDataBundle\DataGrid\ReferenceDataRenderer
+    pim_reference_data.label_renderer.class: Pim\Component\ReferenceData\LabelRenderer
 
 services:
     pim_reference_data.registry:
@@ -23,5 +23,5 @@ services:
         arguments:
             - '@pim_reference_data.repository_resolver'
 
-    pim_reference_data.datagrid_renderer:
-        class: %pim_reference_data.datagrid_renderer.class%
+    pim_reference_data.label_renderer:
+        class: %pim_reference_data.label_renderer.class%
